ChatBot/requirements.txt
absl-py==2.0.0
aiohttp==3.8.5
aiosignal==1.3.1
argcomplete==1.10.3
asttokens==2.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.8.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.0
click==8.1.7
comm==0.1.4
compressed-rtf==1.0.6
contourpy==1.1.1
cycler==0.12.1
dataclasses-json==0.5.14
decorator==5.1.1
docx2txt==0.8
ebcdic==1.1.1
exceptiongroup==1.1.3
executing==2.0.0
extract-msg==0.28.7
faiss-cpu==1.7.4
filelock==3.12.4
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
gast==0.5.4
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.0
grpcio==1.59.0
h5py==3.9.0
huggingface-hub==0.17.3
idna==3.4
IMAPClient==2.1.0
ipython==8.16.1
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyterlab-widgets==3.0.9
keras==2.14.0
kiwisolver==1.4.5
langchain==0.0.150
libclang==16.0.6
lxml==4.9.3
Markdown==3.5
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.7
numpy==1.26.0
oauthlib==3.2.2
olefile==0.46
openai==0.28.1
openapi-schema-pydantic==1.2.4
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pdfminer.six==20191110
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.39
protobuf==4.24.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.19.0
pydantic==1.10.13
Pygments==2.16.1
pyparsing==3.1.1
pypdf==3.16.2
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-pptx==0.6.22
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
six==1.12.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.21
stack-data==0.6.3
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
textract==1.6.5
tiktoken==0.5.1
tokenizers==0.14.1
tqdm==4.66.1
traitlets==5.11.2
transformers==4.34.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
waitress==2.1.2
wcwidth==0.2.8
Werkzeug==3.0.0
widgetsnbextension==4.0.9
wrapt==1.14.1
xlrd==1.2.0
XlsxWriter==3.1.6
yarl==1.9.2
